test/requirements.backup.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anyio
asyncpg
attrs
autocommand
backports.tarfile
certifi
charset-normalizer
click
colorama
dataclasses-json
faiss-cpu
fastapi
filelock
frozenlist
fsspec
greenlet
h11
httpcore
httpx
httpx-sse
huggingface-hub
idna
importlib-metadata
inflect
jaraco.collections
jaraco.context
jaraco.functools
jaraco.text
jinja2
joblib
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
langsmith
mako
markupsafe
marshmallow
more-itertools
mpmath
multidict
mypy-extensions
networkx
numpy
orjson
packaging
pillow
pip
platformdirs
propcache
psycopg2-binary
pydantic
pydantic-core
pydantic-settings
python-dotenv
pytz
pyyaml
regex
requests
requests-toolbelt
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
sniffio
sqlalchemy
starlette
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
tomli
torch
tqdm
transformers
typeguard
typing-extensions
typing-inspect
typing-inspection
unidecode
urllib3
uvicorn
validators
wheel
yarl
zipp
zstandard
